Set ALEX to allow certain words
Previously, ALEX was erroring on a few terms that are required in our
lore, but may be offensive in a different setting (fail{ed,ure},
host{est,sets,s}, and execute). As a quick work around, we had set the
rule to being a warning, but this is unideal since we want it to catch
other potentially offensive language.

Now, Alex allows those words.

Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>